pyproject.toml
[project]
name = "ecom-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "alok-raj",email = "Alok.Raj@mindfiresolutions.com"}
]
readme = "README.md"
requires-python = ">=3.13"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
fastapi = "^0.119.1"
sqlalchemy = "^2.0.44"
psycopg2 = "^2.9.11"
pydantic = {extras = ["email"], version = "^2.12.3"}
python-dotenv = "^1.1.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poe.tasks]
dev = "uvicorn app.main:app --reload"
test = "poetry run python -m unittest discover -s app/tests -p '*.py'"
[dependency-groups]
dev = [
    "poethepoet (>=0.37.0,<0.38.0)"
]

